# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: F1 Draft League API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/by-handle/{handle}:
    get:
      operationId: getUserByHandle
      tags: [users]
      summary: Get user by @handle
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found

  # SEASONS
  /seasons:
    get:
      operationId: listSeasons
      tags: [seasons]
      summary: List all seasons
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Season"

    post:
      operationId: createSeason
      tags: [seasons]
      summary: Create season (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeasonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
...
  /seasons/{id}:
    get:
      operationId: getSeason
      tags: [seasons]
      summary: Get season by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"

    patch:
      operationId: updateSeason
      tags: [seasons]
      summary: Update season (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeasonUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"

  # GRAND PRIX
...
  /constructor-teams:
    get:
      operationId: listConstructorTeams
      tags: [teams]
      summary: List all constructor teams
      parameters:
        - name: seasonId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConstructorTeam"

    post:
      operationId: createConstructorTeam
      tags: [teams]
      summary: Create constructor team (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstructorTeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstructorTeam"
  /constructor-teams/{id}:
    get:
      operationId: getConstructorTeam
      tags: [teams]
      summary: Get constructor team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstructorTeam"

    patch:
      operationId: updateConstructorTeam
      tags: [teams]
      summary: Update constructor team (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstructorTeamUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstructorTeam"

  # DRAFTS (picks per GP)
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, handle, displayName, budget, isAdmin, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        handle:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        budget:
          type: number
        bonusBudget:
          type: number
        isAdmin:
          type: boolean
        createdAt:
          type: string
    UserProfile:
      type: object
      required: [id, handle, displayName, createdAt]
      properties:
        id:
          type: string
        handle:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]

    # SEASONS
    Season:
      type: object
      required: [id, year, name, isActive, baseBudget, createdAt]
      properties:
        id:
          type: string
        year:
          type: integer
        name:
          type: string
        isActive:
          type: boolean
        baseBudget:
          type: number
        createdAt:
          type: string
    SeasonInput:
      type: object
      required: [year, name, baseBudget]
      properties:
        year:
          type: integer
        name:
          type: string
        baseBudget:
          type: number
        isActive:
          type: boolean
    SeasonUpdate:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
        baseBudget:
          type: number

    # GRAND PRIX
...
    ConstructorTeam:
      type: object
      required: [id, seasonId, name, shortName, nationality, price, createdAt]
      properties:
        id:
          type: string
        seasonId:
          type: string
        name:
          type: string
        shortName:
          type: string
        nationality:
          type: string
        color:
          type: ["string", "null"]
        price:
          type: number
        priceChange:
          type: number
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ConstructorTeamInput:
      type: object
      required: [seasonId, name, shortName, nationality, price]
      properties:
        seasonId:
          type: string
        name:
          type: string
        shortName:
          type: string
        nationality:
          type: string
        color:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
    ConstructorTeamUpdate:
      type: object
      properties:
        name:
          type: string
        shortName:
          type: string
        color:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]

    # DRAFTS
...