# lib/api-zod/src/generated/types/constructorTeamUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * F1 Draft League API
 * OpenAPI spec version: 0.1.0
 */

export interface ConstructorTeamUpdate {
  name?: string;
  shortName?: string;
  /** @nullable */
  color?: string | null;
  price?: number;
  /** @nullable */
  imageUrl?: string | null;
}