# lib/api-zod/src/generated/types/seasonUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * F1 Draft League API
 * OpenAPI spec version: 0.1.0
 */

export interface SeasonUpdate {
  name?: string;
  isActive?: boolean;
  baseBudget?: number;
}